move pageTreeCategories to src/pageTree dir
