+++ demo/admin/src/dashboard/LatestContentUpdates.tsx
@@ -2,7 +2,7 @@ import { gql, useQuery } from "@apollo/client";
 import { LatestContentUpdatesDashboardWidget } from "@comet/cms-admin";
 import { useContentScope } from "@src/common/ContentScopeProvider";
 import { GQLLatestContentUpdatesQueryVariables } from "@src/dashboard/LatestContentUpdates.generated";
-import { categoryToUrlParam } from "@src/pages/pageTreeCategories";
+import { categoryToUrlParam } from "@src/pageTree/pageTreeCategories";
 import React from "react";
 
 import { GQLLatestContentUpdatesQuery } from "./LatestContentUpdates.generated";
